7_Lesson_DRF/pyproject.toml
[tool.poetry]
name = "7-lesson-drf"
version = "0.1.0"
description = ""
authors = ["KinChu <chuevgenii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "!=4.2"
djangorestframework = "^3.15.1"
ipython = "^8.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
